through market-history reducer
